Setup: Always set `config.oAuth2` for OAuth2 accounts
